pyproject.toml
[project]
name = "ghpy"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sherpa00", email = "nephack00@gmail.com" }
]
dependencies = [
    "langchain-openai>=0.2.6",
    "python-dotenv>=1.0.1",
    "colorama>=0.4.6",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ghpy"]
